fix: Port ECR Login logic to plugin architecture
